// src/generated/odoc-thin-slice-api.ts
export interface components {
    schemas: {
        SystemInfoResponse: {
            name?: string;
            status?: string;
            /** Format: date-time */
            timestamp?: string;
        };
        ThinSliceCommandRequest: {
            message: string;
        };
        ThinSliceCommandResponse: {
            /** Format: uuid */
            executionId?: string;
            message?: string;
            /** Format: date-time */
            createdAt?: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}